// src/enums.tsx
export enum StarWarsTimeZones {
  GMT = "Coruscant Standard Time",
  GMT_PLUS_1 = "Naboo Time",
  GMT_PLUS_2 = "Tatooine Time",
  GMT_PLUS_3 = "Hoth Time",
  GMT_PLUS_4 = "Alderaan Time",
  GMT_PLUS_5 = "Kashyyyk Time",
  GMT_PLUS_6 = "Dagobah Time",
  GMT_PLUS_7 = "Endor Time",
  GMT_PLUS_8 = "Kamino Time",
  GMT_PLUS_9 = "Mustafar Time",
  GMT_PLUS_10 = "Bespin Time",
  GMT_MINUS_1 = "Lothal Time",
  GMT_MINUS_2 = "Jakku Time",
  GMT_MINUS_3 = "Dathomir Time",
  GMT_MINUS_4 = "Mandalore Time",
  GMT_MINUS_5 = "Yavin 4 Time",
  GMT_MINUS_6 = "Scarif Time",
  GMT_MINUS_7 = "Geonosis Time",
  GMT_MINUS_8 = "Kessel Time",
  GMT_MINUS_9 = "Ord Mantell Time",
  GMT_MINUS_10 = "Ryloth Time",
}